oauthtest/requirements.txt
Django==1.8
django-braces==1.9.0
django-cors-headers==1.2.2
django-oauth-toolkit==0.10.0
djangorestframework==3.2.0
oauthlib==1.0.3
six==1.10.0
